// co-PI-Backend/src/types/enums.ts
export enum Role {
  PI = 'PI',
  CO_INVESTIGATOR = 'CO_INVESTIGATOR',
  ASSISTANT = 'ASSISTANT',
  REVIEWER = 'REVIEWER',
}

export enum ProjectStatus {
  DRAFT = 'DRAFT',
  PROPOSAL_SUBMITTED = 'PROPOSAL_SUBMITTED',
  PROPOSAL_APPROVED = 'PROPOSAL_APPROVED',
  LITERATURE_REVIEW = 'LITERATURE_REVIEW',
  DATA_COLLECTION = 'DATA_COLLECTION',
  DATA_ANALYSIS = 'DATA_ANALYSIS',
  REPORT_WRITING = 'REPORT_WRITING',
  FINAL_SUBMISSION = 'FINAL_SUBMISSION',
  CERTIFICATION = 'CERTIFICATION',
  COMPLETED = 'COMPLETED',
}

export enum EthicalStatus {
  NOT_REQUIRED = 'NOT_REQUIRED',
  PENDING = 'PENDING',
  UNDER_REVIEW = 'UNDER_REVIEW',
  APPROVED = 'APPROVED',
  REJECTED = 'REJECTED',
  EXPIRED = 'EXPIRED',
}

export enum OutputType {
  JOURNAL = 'JOURNAL',
  CONFERENCE = 'CONFERENCE',
  REPORT = 'REPORT',
}